+++ CHANGELOG.md
@@ -24,13 +24,14 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 
-- [#1014](https://gitlab.com/meltano/meltano/issues/1014) Meltano now logs all output from each `meltano elt` run in a log file that uses the unique job_id of the run. It can be found in `.meltano/run/logs/elt_{job_id}.log`.
+- [#1014](https://gitlab.com/meltano/meltano/issues/1014) Meltano now logs all output from each `meltano elt` run in a log file that uses the unique job*id of the run. It can be found in `.meltano/run/logs/elt*{job_id}.log`.
 - [#955](https://gitlab.com/meltano/meltano/issues/955) Establish baseline for demo day and how they should be run
 
 ### Changes
 
 - [#891](https://gitlab.com/meltano/meltano/issues/891) Contributors can run webapp from root directory
 - [#854](https://gitlab.com/meltano/meltano/issues/854) Charts now use pretty labels rather than the ID
+- [#1011](https://gitlab.com/meltano/meltano/issues/1011) Removed "Catch-up Date" in favor of default "Start Date" of extractor
 
 ### Fixes